refactor(tui): remove unused Mermaid library APIs
