--- packages/merman/src/core/terminal/ansi.ts
@@ -1,3 +1,0 @@
-export const ANSI = {
-  reset: "\x1b[0m",
-}
