+++ packages/merman/src/flowchart/options.ts
@@ -1,5 +1,4 @@
 import type { BorderStyle } from "@opentui/core"
-import type { FlowchartDiagramAnsiTheme } from "./style.js"
 import type { FlowchartDirection } from "./types.js"
 
 export interface FlowchartDiagramRenderOptions {
@@ -10,8 +9,4 @@ export interface FlowchartDiagramRenderOptions {
   minRankGap?: number
   /** Fold oversized horizontal layouts vertically when their rendered width exceeds this limit. */
   layoutMaxWidth?: number
-}
-
-export interface FlowchartDiagramAnsiOptions extends FlowchartDiagramRenderOptions {
-  theme?: FlowchartDiagramAnsiTheme
 }
