+++ packages/merman/src/sequence/diagram.ts
@@ -1,12 +1,8 @@
 import { drawSequenceDiagramGrid } from "./drawing.js"
 import { parseMermaidSequenceDiagram } from "./parser.js"
-import { renderSequenceGridAnsi, renderSequenceGridText } from "./render-grid.js"
-import type { SequenceDiagramAnsiOptions, SequenceDiagramRenderOptions } from "./types.js"
+import { renderSequenceGridText } from "./render-grid.js"
+import type { SequenceDiagramRenderOptions } from "./types.js"
 
 export function renderSequenceDiagram(content: string, options: SequenceDiagramRenderOptions = {}): string {
   return renderSequenceGridText(drawSequenceDiagramGrid(parseMermaidSequenceDiagram(content), options))
-}
-
-export function renderSequenceDiagramAnsi(content: string, options: SequenceDiagramAnsiOptions = {}): string {
-  return renderSequenceGridAnsi(drawSequenceDiagramGrid(parseMermaidSequenceDiagram(content), options), options.theme)
 }
