+++ packages/merman/src/sequence/render-grid.ts
@@ -1,13 +1,8 @@
 import type { StyledText } from "@opentui/core"
 import type { DiagramCanvas } from "../core/canvas.js"
-import { renderDiagramGridAnsi, renderDiagramGridStyledText } from "../core/render-grid.js"
-import {
-  resolveSequenceAnsiTheme,
-  sequenceStyleBackgroundColor,
-  sequenceStyleColor,
-  type SequenceStyleColors,
-} from "./style.js"
-import type { SequenceCellStyle, SequenceDiagramAnsiTheme } from "./types.js"
+import { renderDiagramGridStyledText } from "../core/render-grid.js"
+import { sequenceStyleBackgroundColor, sequenceStyleColor, type SequenceStyleColors } from "./style.js"
+import type { SequenceCellStyle } from "./types.js"
 
 export type SequenceGrid = DiagramCanvas<SequenceCellStyle>
 
@@ -24,12 +19,4 @@ export function renderSequenceGridStyledText(
     (run) => sequenceStyleColor(run.style, colors),
     (run) => sequenceStyleBackgroundColor(run.style, colors as Required<SequenceStyleColors>),
   )
-}
-
-export function renderSequenceGridAnsi(grid: SequenceGrid, theme: SequenceDiagramAnsiTheme = {}): string {
-  const resolvedTheme = resolveSequenceAnsiTheme(theme)
-  return renderDiagramGridAnsi(grid, (run) => {
-    if (run.style === "noteBadge") return resolvedTheme.note
-    return run.style ? resolvedTheme[run.style] : undefined
-  })
 }
