+++ packages/merman/src/sequence/style.ts
@@ -1,20 +1,6 @@
 import { RGBA } from "@opentui/core"
-import {
-  ansiBg,
-  ansiFg,
-  createColorRampTheme,
-  createAnsiRampTheme,
-  DIAGRAM_FADE_STEPS,
-  numberedStyleKeys,
-  type DiagramRgb,
-} from "../core/color/style.js"
-import type {
-  AnsiSequenceCellStyle,
-  FadeStyle,
-  MessageStyle,
-  SequenceCellStyle,
-  SequenceDiagramAnsiTheme,
-} from "./types.js"
+import { createColorRampTheme, DIAGRAM_FADE_STEPS, numberedStyleKeys, type DiagramRgb } from "../core/color/style.js"
+import type { FadeStyle, SequenceCellStyle } from "./types.js"
 
 export interface SequenceStyleColors {
   participant?: RGBA
@@ -41,32 +27,6 @@ const DEFAULT_THEME_RGB = {
   noteFg: [215, 229, 221],
   noteBg: [36, 56, 47],
 } as const satisfies Record<string, DiagramRgb>
-
-function createAnsiFadeTheme(style: MessageStyle, from: DiagramRgb, to: DiagramRgb): Record<FadeStyle, string> {
-  return createAnsiRampTheme(numberedStyleKeys(`${style}Fade`, SEQUENCE_FADE_STEPS), from, to) as Record<
-    FadeStyle,
-    string
-  >
-}
-
-const DEFAULT_ANSI_THEME: Required<Record<AnsiSequenceCellStyle, string>> = {
-  participant: ansiFg(DEFAULT_THEME_RGB.participant),
-  lifeline: ansiFg(DEFAULT_THEME_RGB.lifeline),
-  group: ansiFg(DEFAULT_THEME_RGB.group),
-  request: ansiFg(DEFAULT_THEME_RGB.request),
-  response: ansiFg(DEFAULT_THEME_RGB.response),
-  fragment: ansiFg(DEFAULT_THEME_RGB.fragment),
-  fragmentLabel: `${ansiFg(DEFAULT_THEME_RGB.fragment)}${ansiBg(DEFAULT_THEME_RGB.fragmentLabelBg)}`,
-  note: `${ansiFg(DEFAULT_THEME_RGB.noteFg)}${ansiBg(DEFAULT_THEME_RGB.noteBg)}`,
-  ...createAnsiFadeTheme("request", DEFAULT_THEME_RGB.lifeline, DEFAULT_THEME_RGB.request),
-  ...createAnsiFadeTheme("response", DEFAULT_THEME_RGB.lifeline, DEFAULT_THEME_RGB.response),
-}
-
-export function resolveSequenceAnsiTheme(
-  theme: SequenceDiagramAnsiTheme = {},
-): Required<Record<AnsiSequenceCellStyle, string>> {
-  return { ...DEFAULT_ANSI_THEME, ...theme }
-}
 
 export function resolveSequenceStyleColors(
   colors: SequenceStyleColors = {},
