+++ packages/merman/src/state/render-grid.ts
@@ -1,8 +1,8 @@
 import type { StyledText } from "@opentui/core"
 import type { DiagramCanvas } from "../core/canvas.js"
-import { renderDiagramGridAnsi, renderDiagramGridStyledText } from "../core/render-grid.js"
-import { resolveStateAnsiTheme, type StateStyleColors } from "./style.js"
-import type { StateCellStyle, StateDiagramAnsiTheme } from "./types.js"
+import { renderDiagramGridStyledText } from "../core/render-grid.js"
+import type { StateStyleColors } from "./style.js"
+import type { StateCellStyle } from "./types.js"
 
 export type StateGrid = DiagramCanvas<StateCellStyle>
 
@@ -12,13 +12,6 @@ export function renderStateGridText(grid: StateGrid): string {
 
 export function renderStateGridStyledText(grid: StateGrid, colors: StateStyleColors): StyledText {
   return renderDiagramGridStyledText(grid, (run) => (run.style ? colors[run.style] : undefined), undefined, {
-    trimBottom: true,
-  })
-}
-
-export function renderStateGridAnsi(grid: StateGrid, theme: StateDiagramAnsiTheme = {}): string {
-  const resolved = resolveStateAnsiTheme(theme)
-  return renderDiagramGridAnsi(grid, (run) => (run.style ? resolved[run.style] : undefined), {
     trimBottom: true,
   })
 }
