+++ packages/merman/src/test/diagnostics.test.ts
@@ -1,9 +1,13 @@
 import { describe, expect, test } from "bun:test"
-import { Flowchart, MermaidSyntaxError, Sequence, State, render } from "../index.js"
+import { MermaidSyntaxError } from "../diagnostics.js"
+import { parseMermaidFlowchartDiagram } from "../flowchart/parser.js"
+import { parseMermaidSequenceDiagram } from "../sequence/parser.js"
+import { parseMermaidStateDiagram } from "../state/parser.js"
+import { renderSequenceDiagram } from "../sequence/diagram.js"
 
 describe("parser diagnostics", () => {
   test("ignores flowchart presentation directives that do not change terminal structure", () => {
-    const diagram = Flowchart.parse(`flowchart LR
+    const diagram = parseMermaidFlowchartDiagram(`flowchart LR
   A[Start] --> B[Done]
   classDef highlight fill:#fff
   class A highlight
@@ -16,7 +20,7 @@ describe("parser diagnostics", () => {
 
   test("reports unsupported structural flowchart statements with source location", () => {
     expect(() =>
-      Flowchart.parse(`flowchart LR
+      parseMermaidFlowchartDiagram(`flowchart LR
   A[Start] --> B[Done]
   A --- B`),
     ).toThrow('Unsupported syntax in flowchart diagram at line 3: "A --- B"')
@@ -24,12 +28,9 @@ describe("parser diagnostics", () => {
 
   test("exposes structured syntax errors through top-level rendering", () => {
     try {
-      render(
-        `sequenceDiagram
+      renderSequenceDiagram(`sequenceDiagram
   A->>B: request
-  opt retry`,
-        { color: false },
-      )
+  opt retry`)
       throw new Error("expected render to reject unsupported syntax")
     } catch (error) {
       expect(error).toBeInstanceOf(MermaidSyntaxError)
@@ -42,28 +43,28 @@ describe("parser diagnostics", () => {
 
   test("reports unclosed state constructs at their opening line", () => {
     expect(() =>
-      State.parse(`stateDiagram-v2
+      parseMermaidStateDiagram(`stateDiagram-v2
   state Running {
     [*] --> Ready`),
     ).toThrow('Unclosed composite state; expected "}" in state diagram at line 2: "state Running {"')
   })
 
   test("reports unsupported state statements", () => {
-    expect(() => State.parse(`stateDiagram-v2\n  hide empty description`)).toThrow(
+    expect(() => parseMermaidStateDiagram(`stateDiagram-v2\n  hide empty description`)).toThrow(
       'Unsupported syntax in state diagram at line 2: "hide empty description"',
     )
   })
 
   test("reports malformed sequence block endings", () => {
     expect(() =>
-      Sequence.parse(`sequenceDiagram
+      parseMermaidSequenceDiagram(`sequenceDiagram
   end`),
     ).toThrow('Unexpected "end" without an open block in sequence diagram at line 2: "end"')
   })
 
   test("does not attach else through an unclosed nested sequence block", () => {
     expect(() =>
-      Sequence.parse(`sequenceDiagram
+      parseMermaidSequenceDiagram(`sequenceDiagram
   alt available
     loop retry
   else fallback`),
